Fix forever hiding all the content on swipe refresh
Restructure the code a bit to make this happen
